+++ lib/ui/router/router_observer.dart
@@ -11,16 +11,16 @@ class RouterObserver extends _$RouterObserver implements NavigatorObserver {
   String? build() => null;
 
   @override
-  void didPush(Route route, Route? previousRoute) {}
+  void didPush(Route route, Route? previousRoute) => _update(route);
 
   @override
-  void didPop(Route route, Route? previousRoute) {}
+  void didPop(Route route, Route? previousRoute) => _update(route);
 
   @override
-  void didReplace({Route? newRoute, Route? oldRoute}) {}
+  void didReplace({Route? newRoute, Route? oldRoute}) => _update(newRoute);
 
   @override
-  void didRemove(Route route, Route? previousRoute) {}
+  void didRemove(Route route, Route? previousRoute) => _update(route);
 
   @override
   void didChangeTop(Route topRoute, Route? previousTopRoute) =>