+++ test/ui/settings/controls/touch/touch_editor_screen_test.dart
@@ -1,0 +1,15 @@
+import 'package:flutter_test/flutter_test.dart';
+
+import '../../../robot.dart';
+
+void main() {
+  testWidgets('Touch editor screen can be opened', (tester) async {
+    final r = Robot(tester);
+
+    await r.pumpApp();
+    r.settings.showTouchControls = true;
+    await r.mainMenu.tapSettingsButton();
+    await r.settingsScreen.tapControlsTab();
+    await r.settingsScreen.controls.tapTouchEditorButton();
+  });
+}